✅ Make the concurrency tests deterministic and cover two doc'd guarantees
Applies the second code-review round. No Critical or High findings in
either round; these are the Mediums.

Two tests added in the previous commit were flaky by construction. A
caller that *joins* an in-flight fetch never reaches FetchGate, so the
gate cannot observe it — yet both tests opened the gate and then asserted
on coalescing. If the fetch committed first, the joiner would start its
own fetch and the assertion would fail. The same applied to the ten
shared-failure callers, where a straggler finds no cache (failures are
deliberately not memoised) and starts a second fetch. Both now wait on an
entry count the store increments before its first suspension point, which
is a true "has joined" signal. The cancellation test used a fixed yield
loop for the same purpose and now uses the barrier too.

The counter is internal and affects no behaviour; it exists because the
alternative was tests that assert a guarantee they cannot actually
observe.

Also covers two documented guarantees that nothing pinned:

- imagesConfiguration() keeps serving the cached value during an
  in-flight refresh rather than blocking on it. That property comes only
  from the cache check preceding the fetch join, so reordering those two
  lines would silently make every reader wait.
- The idealWidth family, in the live-API suite. It runs a different code
  path from the size: family — it parses the live size strings to pick
  the smallest large enough — so a TMDb-side format change would break it
  without the existing coverage noticing.

Fixes two documentation defects: the changelog still said refresh()
"discards" the cached configuration, contradicting the DocC contract and
the behaviour of the previous commit; and the comment justifying refresh
coalescing had the temporal relationship backwards — the joined fetch was
issued before the joining caller called refresh(), not after. What makes
joining sound is that the fetch is newer than the cached value it
replaces, and the caveat that follows from it is now public.

The store's test file exceeded swiftlint's 400-line and 250-line limits,
so the refresh cases move to their own suite rather than disabling the
rules.
