+++ Tests/TMDbIntegrationTests/ImageIntegrationTests.swift
@@ -55,6 +55,31 @@ struct ImageIntegrationTests {
         #expect(posterURL.absoluteString.contains("/w500/"))
     }
 
+    @Test("posterURL for an ideal width selects a sized variant from the live configuration")
+    func posterURLForIdealWidthSelectsSizedVariant() async throws {
+        let movie = try await client.movies.details(forMovie: 346_698)
+
+        let url = try await client.images.posterURL(for: movie.posterPath, idealWidth: 300)
+
+        // Exercises a different path from the `size:` family: this one parses the
+        // live size strings ("w500", "h632", "original") to pick the smallest
+        // that is at least this wide. A TMDb-side format change would break it
+        // silently, which is what this suite exists to catch.
+        let posterURL = try #require(url)
+        #expect(posterURL.scheme == "https")
+        #expect(!posterURL.absoluteString.contains("/original/"))
+    }
+
+    @Test("profileURL for an ideal width resolves against the live configuration")
+    func profileURLForIdealWidthResolvesAgainstLiveConfiguration() async throws {
+        let person = try await client.people.details(forPerson: 287)
+
+        let url = try await client.images.profileURL(for: person.profilePath, idealWidth: 200)
+
+        let profileURL = try #require(url)
+        #expect(profileURL.scheme == "https")
+    }
+
     @Test("posterURL for a nil path returns nil")
     func posterURLForNilPathReturnsNil() async throws {
         let url = try await client.images.posterURL(for: nil, size: .width(500))
