+++ Tests/TMDbTests/Domain/Services/Images/APIConfigurationStoreTests.swift
@@ -63,9 +63,10 @@ struct APIConfigurationStoreTests {
                 }
             }
 
-            // Hold the window open until a fetch has provably reached the gate,
-            // then release. Without this the test could pass by scheduling luck.
-            await gate.waitUntilEntered(atLeast: 1)
+            // Hold the window open until every caller has provably joined the
+            // in-flight fetch, then release. Without this the test could pass by
+            // scheduling luck.
+            await store.waitUntilCallersEntered(atLeast: 100)
             await gate.open()
 
             var results: [Result<APIConfiguration, TMDbError>] = []
@@ -134,7 +135,11 @@ struct APIConfigurationStoreTests {
                 }
             }
 
-            await gate.waitUntilEntered(atLeast: 1)
+            // All ten must have joined the in-flight fetch before it is released.
+            // A failure is deliberately not memoised, so a straggler arriving
+            // after it completes would find no cache and no in-flight fetch and
+            // start a second one.
+            await store.waitUntilCallersEntered(atLeast: 10)
             await gate.open()
 
             return await group.reduce(into: 0) { count, didFail in
@@ -155,158 +160,6 @@ struct APIConfigurationStoreTests {
         #expect(configurationService.apiConfigurationCallCount == 2)
     }
 
-    @Test("refresh discards the memo, re-fetches, and re-memoises the new value")
-    func refreshDiscardsMemoAndReMemoisesNewValue() async throws {
-        let configurationService = CountingConfigurationService()
-        let firstResult = APIConfiguration.mock(changeKeys: ["first"])
-        let secondResult = APIConfiguration.mock(changeKeys: ["second"])
-        configurationService.enqueue(.success(firstResult))
-        configurationService.enqueue(.success(secondResult))
-        let store = APIConfigurationStore(configurationService: configurationService)
-
-        let first = try await store.apiConfiguration()
-        let cached = try await store.apiConfiguration()
-        #expect(configurationService.apiConfigurationCallCount == 1)
-
-        let refreshed = try await store.refresh()
-        let afterRefresh = try await store.apiConfiguration()
-
-        #expect(first == firstResult)
-        #expect(cached == firstResult)
-        #expect(refreshed == secondResult)
-        #expect(afterRefresh == secondResult)
-        #expect(configurationService.apiConfigurationCallCount == 2)
-    }
-
-    @Test("refresh while a first fetch is in flight does not let the superseded fetch win")
-    func refreshWhileFirstFetchInFlightDoesNotLetSupersededFetchWin() async throws {
-        let gate = FetchGate()
-        let configurationService = CountingConfigurationService(gate: gate)
-        let firstResult = APIConfiguration.mock(changeKeys: ["first"])
-        let secondResult = APIConfiguration.mock(changeKeys: ["second"])
-        configurationService.enqueue(.success(firstResult))
-        configurationService.enqueue(.success(secondResult))
-        let store = APIConfigurationStore(configurationService: configurationService)
-
-        // Park a first fetch at the gate.
-        async let parked = store.apiConfiguration()
-        await gate.waitUntilEntered(atLeast: 1)
-
-        // Supersede it. Its own fetch also parks at the gate.
-        async let refreshed = store.refresh()
-        await gate.waitUntilEntered(atLeast: 2)
-
-        await gate.open()
-
-        let parkedResult = try await parked
-        let refreshedResult = try await refreshed
-
-        // The parked caller asked before the refresh, so it correctly receives the
-        // pre-refresh value; the refresh caller receives the fresh one.
-        #expect(parkedResult == firstResult)
-        #expect(refreshedResult == secondResult)
-        #expect(configurationService.apiConfigurationCallCount == 2)
-
-        // The superseded fetch must not have clobbered the cache, nor detached the
-        // refresh's fetch (which would cause a third fetch here).
-        let afterRefresh = try await store.apiConfiguration()
-
-        #expect(afterRefresh == secondResult)
-        #expect(configurationService.apiConfigurationCallCount == 2)
-    }
-
-    @Test("concurrent refresh calls share a single fetch")
-    func concurrentRefreshCallsShareSingleFetch() async throws {
-        let gate = FetchGate()
-        let configurationService = CountingConfigurationService(gate: gate)
-        let expectedResult = APIConfiguration.mock(changeKeys: ["first"])
-        configurationService.enqueue(.success(expectedResult))
-        configurationService.enqueue(.success(.mock(changeKeys: ["second"])))
-        let store = APIConfigurationStore(configurationService: configurationService)
-
-        let firstRefresh = Task { try await store.refresh() }
-        await gate.waitUntilEntered(atLeast: 1)
-
-        // Arrives while the first refresh is still fetching: it should join that
-        // fetch rather than supersede it, which would waste the first's round trip
-        // and discard its result.
-        let secondRefresh = Task { try await store.refresh() }
-
-        await gate.open()
-
-        let firstResult = try await firstRefresh.value
-        let secondResult = try await secondRefresh.value
-
-        #expect(firstResult == expectedResult)
-        #expect(secondResult == expectedResult)
-        #expect(configurationService.apiConfigurationCallCount == 1)
-
-        // …and that shared fetch must still have been memoised.
-        let cached = try await store.apiConfiguration()
-
-        #expect(cached == expectedResult)
-        #expect(configurationService.apiConfigurationCallCount == 1)
-    }
-
-    @Test("a failed refresh keeps the previously cached configuration")
-    func failedRefreshKeepsPreviouslyCachedConfiguration() async throws {
-        let configurationService = CountingConfigurationService()
-        let expectedResult = APIConfiguration.mock(changeKeys: ["first"])
-        configurationService.enqueue(.success(expectedResult))
-        configurationService.enqueue(.failure(.unknown))
-        let store = APIConfigurationStore(configurationService: configurationService)
-
-        let before = try await store.apiConfiguration()
-        #expect(configurationService.apiConfigurationCallCount == 1)
-
-        await #expect(throws: TMDbError.unknown) {
-            _ = try await store.refresh()
-        }
-
-        // A refresh that fails must not destroy a known-good configuration: losing
-        // it would make every later image URL re-fetch until the network recovers.
-        let after = try await store.apiConfiguration()
-
-        #expect(before == expectedResult)
-        #expect(after == expectedResult)
-        #expect(configurationService.apiConfigurationCallCount == 2)
-    }
-
-    @Test("a superseded fetch that fails does not detach the refresh's fetch")
-    func supersededFetchThatFailsDoesNotDetachRefreshFetch() async throws {
-        let gate = FetchGate()
-        let configurationService = CountingConfigurationService(gate: gate)
-        let expectedResult = APIConfiguration.mock(changeKeys: ["second"])
-        configurationService.enqueue(.failure(.unknown))
-        configurationService.enqueue(.success(expectedResult))
-        let store = APIConfigurationStore(configurationService: configurationService)
-
-        // Explicit `Task` handles rather than `async let`: an `async let` binding
-        // cannot be captured by the `#expect(throws:)` closure.
-        let parkedCaller = Task { try await store.apiConfiguration() }
-        await gate.waitUntilEntered(atLeast: 1)
-
-        let refreshCaller = Task { try await store.refresh() }
-        await gate.waitUntilEntered(atLeast: 2)
-
-        await gate.open()
-
-        await #expect(throws: TMDbError.unknown) {
-            _ = try await parkedCaller.value
-        }
-        let refreshedResult = try await refreshCaller.value
-
-        #expect(refreshedResult == expectedResult)
-        #expect(configurationService.apiConfigurationCallCount == 2)
-
-        // The superseded failure must not have cleared the refresh's committed
-        // cache entry — a third fetch here would mean it did.
-        let afterRefresh = try await store.apiConfiguration()
-
-        #expect(afterRefresh == expectedResult)
-        #expect(configurationService.apiConfigurationCallCount == 2)
-    }
-
     @Test("cancelling one caller does not starve the others sharing the fetch")
     func cancellingOneCallerDoesNotStarveTheOthers() async throws {
         let gate = FetchGate()
@@ -318,17 +171,12 @@ struct APIConfigurationStoreTests {
         let cancelledCaller = Task { try await store.apiConfiguration() }
         await gate.waitUntilEntered(atLeast: 1)
 
-        // These two are given a chance to join the in-flight fetch before the gate
-        // opens. Their ordering is not guaranteed — if they arrive late they read
-        // the committed cache instead — so they are corroborating, not
-        // load-bearing. The guarantee this test actually rests on is the cancelled
-        // caller below, which `waitUntilEntered` proves was parked inside the
-        // shared fetch at the moment it was cancelled.
+        // Both must have joined the same in-flight fetch before the gate opens,
+        // otherwise they would read the committed cache instead and the test
+        // would not exercise sharing at all.
         let second = Task { try await store.apiConfiguration() }
         let third = Task { try await store.apiConfiguration() }
-        for _ in 0 ..< 10 {
-            await Task.yield()
-        }
+        await store.waitUntilCallersEntered(atLeast: 3)
 
         cancelledCaller.cancel()
         await gate.open()
